+++ docs/CHANGELOG.md
@@ -24,6 +24,7 @@ There are some breaking changes:
 - The new way to set up `AgentType` instances with an associated `Market` is to create them (with the agents in the `Market`'s `agents` attribute), then invoke the `Market`'s new `give_agent_params()` method. #1719
 - The above method calls each `agent`'s `get_market_params()` method, which references the `market_vars` class attribute for the names of objects to take from the associated `Market`.
 - All interpolator classes now have default derivative methods using finite differences. These are fallback methods, and are already overridden by most subclasses. #1723
+- New consumption-saving model with habit formation has been added; extends IndShockConsumerType model. #1739
 
 #### Minor Changes